Refactor for better TclObj support
Refactors a substantial portion of the existing library to allow
for custom objects and types (still WIP).  In particular, it pulls
out the set of functions needed to interact with objects inside of
their callbacks where the interpreter is not available and makes
them lazily initialized globals.  The default implementations for
the reference counting are included because 8.6.10 does not have
these initialized in the stub.